Remote and local video streams both showing.
