+++ vision/src/test/redux.test.ts
@@ -1,5 +1,7 @@
 import {createStore} from "redux";
 import WebSocket from "ws";
+import { BehaviorSubject } from "rxjs";
+import { combineLatest, map } from "rxjs/operators";
 
 import store from "../state/store";
 import {logger} from "../logger";
